requirements.txt
progress
elasticsearch>=2.0.0,<3.0.0
PyMySQL
pyyaml
pymongo
